Renaming loader to historic_data_loader
